:recycle: merge CollectionViewCell
